Text.CSV

Description

This is a library for parsing comma separated values (CSV).

Documentation

parseCSV :: String -> [[String]]

The parseCSV function parses a String containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. Note that CSV is ambiguous, in that a newline may be intended either as a record terminator, or as a record separator. The parseCSV function treats newlines as record separators, which allows it to successfully parse in either case, albeit with a spurious final record consisting of a single, empty field, for CSV files in which newline is a record terminator.

Client code should be aware of the possibility (even likelihood) of encountering records that consist of a single, empty field.

readCSV :: FilePath -> IO [[String]]

The readCSV function parses a file containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. The same caveats regarding CSV that are noted in documentation for parseCSV apply to this function as well.