Text.CSV

Description

A module for handling comma-separated values (CSV), either in Strings or files.

Documentation

parseCSV :: String -> [[String]]

The parseCSV function parses a String containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. This function, unlike parseRawCSV, filters out trivial records, i.e., records that consist of a single, white-space only field.

readCSV :: FilePath -> IO [[String]]

The readCSV function parses a file containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. This function uses parseCSV for parsing.

parseRawCSV :: String -> [[String]]

The parseRawCSV function parses a String containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. Note that CSV specification is ambiguous, in that a terminal newline may be intended either as a record terminator, or as a record separator. The parseRawCSV function treats newlines as record separators, which allows it to successfully parse in either case, albeit with a spurious final record consisting of a single, empty field, given an argument String in which newline was intended as a record terminator.

Client code should be aware of the possibility (even likelihood) of encountering trivial records.

readRawCSV :: FilePath -> IO [[String]]

The readRawCSV function parses a file containing comma separated values (CSV) as a list of records, each of which is a list of fields of type String. This function uses parseRawCSV for parsing, and so may contain trivial records.

prune :: [[String]] -> [[String]]

The prune function is used to filter the values obtained by the application of parseCSV or readCSV, eliminating any trivial records, i.e., records that consist of a single field that consists only of white space.

encodeCSVField :: String -> String

The encodeCSVField function will quote a String that contains carriage returns, linefeeds, double-quotes, or commas, properly quoting for CSV any double-quoted fields.

showCSV :: [[String]] -> String

The showCSV function converts a [[String]] to a String, following the conventions of CSV, i.e., the result string will be valid CSV. The intent is that

      readCSV . showCSV = id

This may not be precisely so if the original CSV contains trivial records, as they will be elided by the process.

writeCSV :: FilePath -> [[String]] -> IO ()

The writeCSV function takes a [[String]], converting it to what is intended to be equivalent CSV (note the caveat in showCSV), and writing the results to the indicated file.